fix type from int into double

+++ Second_GitHub_Calculator_CW12/Program.cs
@@ -7,15 +7,15 @@
             Console.WriteLine("Calculator\n");
 
             Console.WriteLine("Enter Number 1:");
-            int number1 = int.Parse(Console.ReadLine());
+            double number1 = double.Parse(Console.ReadLine());
 
             Console.WriteLine("Select a Process(+,-,*,/,%):");
             string process = Console.ReadLine();
 
             Console.WriteLine("Enter Number 2:");
-            int number2 = int.Parse(Console.ReadLine());
+            double number2 = double.Parse(Console.ReadLine());
 
-            int result = 0;
+            double result = 0;
             switch (process)
             {
